Clean up utils
