+++ src/utils/index.ts
@@ -1,3 +1,5 @@
 export * from "./dolphin";
 export * from "./slpFile";
 export * from "./controls";
+export * from "./winner";
+export * from "./helpers";
