+++ test/realtime.spec.ts
@@ -1,7 +1,6 @@
 import sinon from "sinon";
 
-import { SlpRealTime, SlpStream } from "../src";
-import { pipeFileContents } from "./helpers";
+import { pipeFileContents, SlpRealTime, SlpStream } from "../src";
 import { Subscription } from "rxjs";
 
 describe("SlpRealTime", () => {
